refactor: 패키지 구조 변경 (#25) - 패키지 구조 변경 - application.yml 파일 내 active로 변경 - data.sql 파일로 더미 데이터 관리
