--- src/main/java/com/develop_ping/union/gathering/controller/dto/GatheringResponse.java
@@ -1,4 +1,0 @@
-package com.develop_ping.union.gathering.controller.dto;
-
-public class GatheringResponse {
-}
