+++ src/main/java/com/develop_ping/union/gathering/domain/GatheringManager.java
@@ -1,6 +1,7 @@
 package com.develop_ping.union.gathering.domain;
 
-import com.develop_ping.union.gathering.application.dto.GatheringInfo;
+import com.develop_ping.union.gathering.domain.dto.GatheringInfo;
+import com.develop_ping.union.gathering.domain.entity.Gathering;
 
 // 추상화 요놈이 infra layer 이름만 store다.
 public interface GatheringManager {
