+++ src/main/java/com/develop_ping/union/gathering/infra/GatheringManagerImpl.java
@@ -1,7 +1,7 @@
 package com.develop_ping.union.gathering.infra;
 
-import com.develop_ping.union.gathering.application.dto.GatheringInfo;
-import com.develop_ping.union.gathering.domain.Gathering;
+import com.develop_ping.union.gathering.domain.dto.GatheringInfo;
+import com.develop_ping.union.gathering.domain.entity.Gathering;
 import com.develop_ping.union.gathering.domain.GatheringManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
