+++ src/main/java/com/develop_ping/union/gathering/infra/GatheringRepository.java
@@ -1,6 +1,6 @@
 package com.develop_ping.union.gathering.infra;
 
-import com.develop_ping.union.gathering.domain.Gathering;
+import com.develop_ping.union.gathering.domain.entity.Gathering;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
