+++ src/main/java/com/develop_ping/union/gathering/presentation/dto/GatheringResponse.java
@@ -1,0 +1,4 @@
+package com.develop_ping.union.gathering.presentation.dto;
+
+public class GatheringResponse {
+}
